+++ lib/cisco_node_utils/cisco_cmn_utils.rb
@@ -113,6 +113,14 @@ module Cisco
       return true if Platform.chassis['pid'][ver_regexp]
     end
 
+    def self.fretta?
+      require_relative 'platform'
+      Platform.slots.each do |_x, row|
+        return true if row['pid'][/-R/]
+      end
+      false
+    end
+
     # Helper utility method for ip/prefix format networks.
     # For ip/prefix format '1.1.1.1/24' or '2000:123:38::34/64',
     # we need to mask the address using the prefix length so that they
@@ -362,9 +370,9 @@ module Cisco
     end # merge_range
 
     def self.add_quotes(value)
-      return value if image_version?(/7.3/)
-      value = "\"#{value}\"" unless
-        value.start_with?('"') && value.end_with?('"')
+      return value if image_version?(/7.3.[0-1]/) || value.nil?
+      value = "'#{value}'" unless
+        value.start_with?('"', "'") && value.end_with?('"', "'")
       value
     end # add_quotes