+++ tests/test_interface_switchport.rb
@@ -80,11 +80,7 @@ class TestSwitchport < TestInterfaceSwitchport
 
   def test_access_vlan_sw_disabled
     interface.switchport_mode = :disabled
-    if platform == :ios_xr
-      assert_nil(interface.access_vlan)
-    else
-      assert_equal(DEFAULT_IF_ACCESS_VLAN, interface.access_vlan)
-    end
+    assert_nil(interface.access_vlan)
   end
 
   def test_access_vlan_sw_trunk
@@ -414,12 +410,13 @@ class TestInterfaceSwVtp < TestInterfaceSwitchport
 
   def setup
     super
-    skip('VTP is not supported on IOS XR') if platform == :ios_xr
+    skip('VTP is not supported on IOS XR or fretta') if
+      platform == :ios_xr || product_tag[/n(3|9)k-f/]
     @vtp = Vtp.new(true)
   end
 
   def teardown
-    vtp.destroy unless platform == :ios_xr
+    vtp.destroy unless platform == :ios_xr || product_tag[/n(3|9)k-f/]
     super
   end